starter/requirements.txt
a2a-sdk>=0.3.6
google-adk>=1.17.0
toolbox-core>=0.5.0
google-cloud-storage>=2.16.0
pdfminer.six>=20231228
